pjt-final-drf/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.5
django-cors-headers==3.7.0
django-extensions==3.1.5
django-heroku==0.3.1
django-imagekit==4.1.0
django-modeltranslation==0.16.1
django-taggit==2.0.0
django-translator==1.0.15
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
pilkit==2.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.6
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
